pyproject.toml
[project]
name = "bert4ner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "certifi==2025.7.14",
    "charset-normalizer==3.4.2",
    "colorama==0.4.6",
    "filelock==3.18.0",
    "fsspec==2025.7.0",
    "hf-xet==1.1.5",
    "huggingface-hub==0.33.4",
    "idna==3.10",
    "jinja2==3.1.6",
    "joblib==1.5.1",
    "markupsafe==3.0.2",
    "mpmath==1.3.0",
    "networkx==3.4.2",
    "numpy==1.26.4",
    "packaging==25.0",
    "pytorch-crf==0.7.2",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.4",
    "safetensors==0.5.3",
    "scikit-learn==1.7.0",
    "scipy==1.15.3",
    "seqeval==1.2.2",
    "sympy==1.14.0",
    "threadpoolctl==3.6.0",
    "tokenizers==0.21.2",
    "torch==2.7.1",
    "tqdm==4.67.1",
    "transformers==4.53.2",
    "typing-extensions==4.14.1",
    "urllib3==2.5.0",
]
